rename pagination component files
