+++ src/components/PageSections/Pagination/Pagination.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Pagination = ({
+  pagesArray,
+  PageComponent,
+  filterable=false,
+  searchedTerm,
+  smallMargins=false,
+}) => {
+  return (
+
+  )
+}
