+++ src/components/Pages/News/NewsList.js
@@ -5,7 +5,7 @@ import { news as i18n } from 'js/i18n/definitions';
 import PageHeader from 'components/PageHeader';
 
 import ContextualNav from 'components/PageSections/ContextualNav';
-import PaginationContainer from 'components/PageSections/Pagination/paginationContainer.js';
+import PaginationContainer from 'components/PageSections/Pagination/PaginationContainer.js';
 import UserFeedback from 'components/UserFeedback';
 
 const NewsListItem = ({ page }) => {
